// client/src/types/Handout.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { HandoutVisibility } from "./HandoutVisibility";

export type Handout = { id: string, campaign_id: string, title: string, content: string, visibility: HandoutVisibility, player_ids: Array<string>, created_by: string, created_at: string, updated_at: string, };